Fix #11205 - Enable Compiler plugin Test cases
