+++ tests/ballerina-compiler-plugin-test/src/test/java/io/ballerina/test/compiler/plugins/TestEvent.java
@@ -27,15 +27,10 @@ class TestEvent {
     enum Kind {
         PKG_NODE("package"),
         SERVICE_ANN("service"),
-        RESOURCE_ANN("resource"),
-        CONNECTOR_ANN("connector"),
-        ACTION_ANN("action"),
-        STRUCT_ANN("struct"),
         TYPEDEF_ANN("struct"),
         FUNC_ANN("function"),
         VARIAVLE_ANN("variable"),
         ANNOTATION_ANN("annotation"),
-        TRANSFORM_ANN("transformer"),
         CODE_GEN("codegen");
 
         String name;
