+++ tests/ballerina-compiler-plugin-test/src/test/resources/META-INF/services/org.ballerinalang.compiler.plugins.CompilerPlugin
@@ -1,2 +1,2 @@
-io.ballerina.test.compiler.plugins.ABCCompilerPlugin
-io.ballerina.test.compiler.plugins.XYZCompilerPlugin
+io.ballerina.test.compiler.plugins.TestCompilerPlugin
+io.ballerina.test.compiler.plugins.FunctionsTestCompilerPlugin
