--- tests/ballerina-compiler-plugin-test/src/test/resources/test-src/compiler_plugin/formats/yaml/yaml.bal
@@ -1,4 +1,0 @@
-
-public annotation YAMLConfig attach function, action {
-    string value;
-}
