--- tests/ballerina-compiler-plugin-test/src/test/resources/test-src/compiler_plugin/images/png/ws.bal
@@ -1,12 +1,0 @@
-
-public annotation IConfig attach resource, function, action {
-    string value;
-}
-
-public annotation IData attach struct, service, connector {
-    string value;
-}
-
-public annotation Element attach const, annotation, transformer {
-    string value;
-}
