+++ tests/ballerina-compiler-plugin-test/src/test/resources/test-src/test-project/Ballerina.toml
@@ -1,0 +1,5 @@
+[project]
+org-name= "testOrg"
+version= "1.0.0"
+
+[dependencies]
