+++ tests/ballerina-compiler-plugin-test/src/test/resources/test-src/test-project/src/services/data.bal
@@ -1,0 +1,2 @@
+
+public annotation record { string value;} Config on resource function, service;
